🐞DbConfig.idType

+++ gk5-test/gk5-test-mybatis-x/src/main/java/wang/bannong/gk5/test/mybatisx/common/ShiroUser.java
@@ -8,10 +8,10 @@ import lombok.Data;
 public class ShiroUser implements Serializable {
     private static final long serialVersionUID = 4170935004841605924L;
 
-    private Integer id;
-    private String  name;
-    private String  role;
-    private String  passwd;
-    private String  permission;
-    private Long    score;
+    private Long   id;
+    private String name;
+    private String role;
+    private String passwd;
+    private String permission;
+    private Long   score;
 }
